Make Queue own either EmptyQueue or ListQueue
